add quick menu

+++ install.sh
@@ -67,7 +67,9 @@ $sudo rm -f ~/.vim/bundle/vim-snipmate/snippets/*
 cp vim-snipmate/*.snippets ~/.vim/bundle/vim-snipmate/snippets/
 
 if [ -d /usr/local/lib/ ]; then
-    sudo cp vim/javaPlugin/google-java-format-1.4-all-deps.jar /usr/local/lib/
+    if [ ! -f /usr/local/lib/google-java-format-1.4-all-deps.jar ]; then
+        sudo cp vim/javaPlugin/google-java-format-1.4-all-deps.jar /usr/local/lib/
+    fi
 fi
